requirements.txt
fastapi==0.73.0
uvicorn==0.17.4
pydantic[email]
python-decouple==3.8
PyJWT==2.5.0
pymongo==4.3.3
passlib==1.7.4